requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.11.28
coverage==4.5.4
docopt==0.6.2
Hexy==1.4.3
importlib-metadata==0.19
mock==3.0.5
more-itertools==7.2.0
OpenEye-toolkits==2019.10.2
OpenEye-toolkits-python3-osx-x64==2019.10.2
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==4.6.5
pytest-cov==2.7.1
setuptools-scm==3.3.3
six==1.12.0
wcwidth==0.1.7
zipp==0.6.0
